Time: 159 ms (74.60%), Space: 103 MB (9.52%) - LeetHub

+++ 2110-number-of-smooth-descent-periods-of-a-stock/2110-number-of-smooth-descent-periods-of-a-stock.cpp
@@ -3,7 +3,7 @@ public:
     long long getDescentPeriods(vector<int>& prices) {
         int n = prices.size();
         long long ans = 0;
-        long long cur, prev = 1;
+        int cur, prev = 1;
         for (int i=0; i<n; ++i) {
             if (i>0 && prices[i-1]-1 == prices[i]) {
                 cur = prev+1;
